feat: cd: Add release pipeline
A GHA pipeline is created for publishing draft releases on tag push.

The main commands (archive and checksum generation) are added to the
helper `exec` script.

Cargo's release profile is customized to decrease the binary size as
much as possible without disrupting how `fcb` works.

Last but not least, additional metadata is added to the manifest file.

Signed-off-by: acikgozb <acikgozb@proton.me>